+++ Runtime/Models/ECommerceModels.cs
@@ -410,6 +410,26 @@ namespace AccelByte.Models
         [DataMember] public int? limit { get; set; }
         [DataMember] public string sortBy { get; set; }
     }
+    
+    [DataContract]
+    public class ItemCriteriaV2
+    {
+        [DataMember] public string storeId { get; set; }
+        [DataMember] public string categoryPath { get; set; }
+        [DataMember] public bool includeSubCategoryItem { get; set; } = false;// Default == false. Can change to "true". Anything != "true" will remain default
+        [DataMember] public ItemType itemType { get; set; }
+        [DataMember] public EntitlementAppType appType { get; set; }
+        [DataMember] public string baseAppId { get; set; }
+        [DataMember] public string[] tags { get; set; }
+        [DataMember] public string[] features { get; set; }
+        [DataMember] public bool activeOnly { get; set; }
+        [DataMember] public string region { get; set; }
+        [DataMember] public DateTime availableDate { get; set; }
+        [DataMember] public string targetNamespace { get; set; }
+        [DataMember] public int? offset { get; set; }
+        [DataMember] public int? limit { get; set; }
+        [DataMember] public string sortBy { get; set; }
+    }
 
     [DataContract]
     public class Image
@@ -592,6 +612,57 @@ namespace AccelByte.Models
     public class ItemPagingSlicedResult
     {
         [DataMember] public ItemInfo[] data { get; set; }
+        [DataMember] public Paging paging { get; set; }
+    }
+
+    [DataContract]
+    public class ItemInfoV2
+    {
+        [DataMember] public string itemId { get; set; }
+        [DataMember] public string appId { get; set; }
+        [DataMember] public EntitlementAppType appType { get; set; } //"GAME" 
+        [DataMember] public SeasonType SeasonType { get; set; }
+        [DataMember] public string baseAppId { get; set; }
+        [DataMember] public string sku { get; set; }
+        [DataMember] public string name { get; set; }
+        [DataMember] public EntitlementType entitlementType { get; set; }
+        [DataMember] public int useCount { get; set; }
+        [DataMember] public bool stackable { get; set; }
+        [DataMember] public string categoryPath { get; set; }
+        [DataMember] public Image[] images { get; set; }
+        [DataMember] public string thumbnailUrl { get; set; }
+        [DataMember] public Dictionary<string, object> localizations { get; set; }
+        [DataMember] public ItemStatus status { get; set; }
+        [DataMember] public bool listable { get; set; }
+        [DataMember] public bool purchasable { get; set; }
+        [DataMember] public ItemType itemType { get; set; }
+        [DataMember] public string targetNamespace { get; set; }
+        [DataMember] public string targetCurrencyCode { get; set; }
+        [DataMember] public string targetItemId { get; set; }
+        [DataMember] public Dictionary<string, object> regionData { get; set; }
+        [DataMember] public ItemRecurring recurring { get; set; }
+        [DataMember] public string[] itemIds { get; set; }
+        [DataMember] public Dictionary<string, int> itemQty { get; set; }
+        [DataMember] public string[] boundItemIds { get; set; }
+        [DataMember] public string[] tags { get; set; }
+        [DataMember] public string[] features { get; set; }
+        [DataMember] public int maxCountPerUser { get; set; }
+        [DataMember] public int maxCount { get; set; }
+        [DataMember] public string clazz { get; set; }
+        [DataMember] public Dictionary<string, object> ext { get; set; }
+        [DataMember] public string boothName { get; set; }
+        [DataMember] public int displayOrder { get; set; }
+        [DataMember] public DateTime createdAt { get; set; }
+        [DataMember] public DateTime updatedAt { get; set; }
+        [DataMember] public ItemPurchaseCondition purchaseCondition { get; set; }
+        [DataMember] public ItemOptionBoxConfig optionBoxConfig { get; set; }
+        [DataMember] public LootBoxConfig lootBoxConfig { get; set; }
+    }
+    
+    [DataContract]
+    public class ItemPagingSlicedResultV2
+    {
+        [DataMember] public ItemInfoV2[] data { get; set; }
         [DataMember] public Paging paging { get; set; }
     }
 
@@ -942,6 +1013,9 @@ namespace AccelByte.Models
     [DataContract]
     public class PlayStationDLCSync
     {
+        [DataMember] public string productId { get; set; }
+        [DataMember] public int price { get; set; }
+        [DataMember] public string currencyCode { get; set; }
         [DataMember] public int serviceLabel { get; set; }
     }